% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aMAXtheory.R
\name{aMAXtheory}
\alias{aMAXtheory}
\title{Function to compute stationary distribution of HMM}
\usage{
aMAXtheory(amin, K, ref = c("Rousseau", "Gassiat"))
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

