% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aminTheory.R
\name{aminTheory}
\alias{aminTheory}
\title{Function to compute stationary distribution of HMM}
\usage{
aminTheory(amin, K)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

