% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dDirichlet}
\alias{dDirichlet}
\title{A dDirichlet Function}
\usage{
dDirichlet(x, alpha, log = FALSE)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

