% Generated by roxygen2 (4.0.2): do not edit by hand
\name{density_f2}
\alias{density_f2}
\title{Function to compute, at 1 iteration,}
\usage{
density_f2(y, .mu, .Q, .q0)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

