% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/formu2.R
\name{formu2}
\alias{formu2}
\title{Funkme_MetHast_Q}
\usage{
formu2(stateChain, Y, K = K)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
text...
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

