% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getq0_trycatch}
\alias{getq0_trycatch}
\title{Function to compute stationary distribution of HMM}
\usage{
getq0_trycatch(Q)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

