% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getq0NEW.R
\name{getq0NEW}
\alias{getq0NEW}
\title{Function to compute stationary distribution of HMM}
\usage{
getq0NEW(Q)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
#
}
\keyword{dirichlet}

