% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gibbsHMM.R
\name{gibbsHMM}
\alias{gibbsHMM}
\title{gibbsHMM}
\usage{
gibbsHMM(YZ, M = 2000, K = 5, mu0 = 0, var0 = 100, alphaMin = 0.5,
  p = 1)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

