% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gibbsHMM_2}
\alias{gibbsHMM_2}
\title{gibbsHMM_PT}
\usage{
gibbsHMM_2(YZ, M = 2000, K = 10, alphaMAX = 1, alphaMin = 1e-30,
  lab = "sim", type = "First")
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

