% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gibbsHMM_LYRA2.R
\name{gibbsHMM_LYRA2}
\alias{gibbsHMM_LYRA2}
\title{gibbsHMM_PT}
\usage{
gibbsHMM_LYRA2(YZ, densTrue, M = 2000, burn = 500, type = 1,
  alphaMAX = 1, alphaMin = 0.001, J = 20)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
parallel tempering with a column prior - option to mix over column or stick to j=1
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

