% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gibbsHMM_PT_FlexIt.R
\name{gibbsHMM_PT_FlexIt}
\alias{gibbsHMM_PT_FlexIt}
\title{gibbsHMM_PT}
\usage{
gibbsHMM_PT_FlexIt(YZ, M = 2000, K = 10, mu0 = 0, var0 = 100,
  alphaMAX = 1, alphaMin = 1e-10, J = 20, lab = "sim", type = "First",
  InitType = c("PropObs", "UniIni"))
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
HMM gibbs with PT, initial dist not stationary dist
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

