% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gibbsHMM_PTnew}
\alias{gibbsHMM_PTnew}
\title{gibbsHMM_PT}
\usage{
gibbsHMM_PTnew(YZ, M = 2000, K = 10, alphaMAX = 1, type = c(1, "mix"),
  alphaMin = 1e-30, J = 20, lab = "sim")
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
parallel tempering with a column prior - option to mix over column or stick to j=1
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

