% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gibbsHMM_PTnew_Mixture.R
\name{gibbsHMM_PTnew_Mixture}
\alias{gibbsHMM_PTnew_Mixture}
\title{gibbsHMM_PT}
\usage{
gibbsHMM_PTnew_Mixture(YZ, M = 2000, K = 10, alphaMAX = 1,
  PrbDiag = c("half", "fair"), alphaMin = 1e-30, J = 20, lab = "sim")
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

