% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gibbsHMM_SimpleMix.R
\name{gibbsHMM_SimpleMix}
\alias{gibbsHMM_SimpleMix}
\title{gibbsHMM_PT}
\usage{
gibbsHMM_SimpleMix(YZ, M = 2000, K = 10, alphaMAX = 1, alphaMin = 0.001,
  PrbDiag = c("half", "fair"), lab = "sim")
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

