% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GibbsHMMbeta_manual}
\alias{GibbsHMMbeta_manual}
\title{GibbsHMMbeta_manual}
\usage{
GibbsHMMbeta_manual(Y, M, alpha1 = aMax, alpha2 = alphaMin,
  beta1 = alphaMin, beta2 = aMax)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

