% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gibbsHMMdiag}
\alias{gibbsHMMdiag}
\title{gibbsHMMdiag}
\usage{
gibbsHMMdiag(YZ, M = 2000, K = 5, mu0 = 0, var0 = 100, alphaMin = 0.5,
  p = 1)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

