% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HmmAllocationPlot.R
\name{HmmAllocationPlot}
\alias{HmmAllocationPlot}
\title{PostProcessing function univ}
\usage{
HmmAllocationPlot(outZ, myY)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

