% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeStartSimpler.R
\name{makeStartSimpler}
\alias{makeStartSimpler}
\title{gibbsHMM}
\usage{
makeStartSimpler(Y, k = K)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

