% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maxZ.R
\name{maxZ}
\alias{maxZ}
\title{Compute median state from unswitched MCMC samples from Zmix}
\usage{
maxZ(x)
}
\arguments{
\item{x}{}
}
\description{
Compute median state
}
\examples{
#something
}
\keyword{states}

