% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MuXqoPlot.R
\name{MuXqoPlot}
\alias{MuXqoPlot}
\title{MuXqoPlot}
\usage{
MuXqoPlot(gibbsout, burn, trueValues = "NA", minq = 0.01, plotlab = "")
}
\arguments{
\item{...}{}
}
\description{
Plot means by stationary distribution
}
\examples{
...
}
\keyword{...}

