% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MuXqoPlotALL.R
\name{MuXqoPlotALL}
\alias{MuXqoPlotALL}
\title{MuXqoPlot}
\usage{
MuXqoPlotALL(gibbsout, burn, trueValues = NA, plotlab = "")
}
\arguments{
\item{...}{}
}
\description{
Plot means by stationary distribution
}
\examples{
...
}
\keyword{...}

