% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallelAccept.R
\name{parallelAccept}
\alias{parallelAccept}
\title{UpdateStates}
\usage{
parallelAccept(w1, w2, a1, a2)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

