% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parallelAcceptHMM_Aug15}
\alias{parallelAcceptHMM_Aug15}
\title{Prior Parallel Tempering acceptance ratio}
\usage{
parallelAcceptHMM_Aug15(Qchain1, Qchain2, Alpha1, Alpha2, X1all, X2all)
}
\arguments{
\item{Qchain1,}{Qchain2, Alpha1, Alpha2,  X1all, X2all}
}
\description{
acceptance ratio
}
\examples{
#TBA
}
\keyword{dirichlet}

