% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallelAcceptHMM_ergodic.R
\name{parallelAcceptHMM_ergodic}
\alias{parallelAcceptHMM_ergodic}
\title{UpdateStates}
\usage{
parallelAcceptHMM_ergodic(Qchain1, Qchain2, Alpha1, Alpha2)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

