% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parallelAcceptHMM_QV}
\alias{parallelAcceptHMM_QV}
\title{UpdateStates}
\usage{
parallelAcceptHMM_QV(Qchain1, Qchain2, Alpha1, Alpha2, X1, X2)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

