% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallelAcceptHMM_QV2.R
\name{parallelAcceptHMM_QV2}
\alias{parallelAcceptHMM_QV2}
\title{UpdateStates}
\usage{
parallelAcceptHMM_QV2(Qchain1, Qchain2, Alpha1, Alpha2, X1, X2)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
acceptance ratio based on transition matrix and initial dist
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

