% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/processHMMs.R
\name{processHMMs}
\alias{processHMMs}
\title{processHMMs}
\usage{
processHMMs(Result, K = 5, plotname = "", simMeans = c(5, 2, 8))
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

