% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReplicateSimer}
\alias{ReplicateSimer}
\title{gibbsHMM_PT}
\usage{
ReplicateSimer(N, n, SimID, ITERATIONS, BURN, AMAX, PRIOR_TYPE, PTchain = 20)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
parallel tempering with a column prior - option to mix over column or stick to j=1
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

