% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trimit_HMM.R
\name{trimit_HMM}
\alias{trimit_HMM}
\title{gibbsHMM}
\usage{
trimit_HMM(Out = Out, nEnd = EndSize)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

