% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TrimThin.R
\name{TrimThin}
\alias{TrimThin}
\title{Trim n Thin}
\usage{
TrimThin(out, burn, Thin)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

