% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/UpdateStatesNewINIT.R
\name{UpdateStatesNewINIT}
\alias{UpdateStatesNewINIT}
\title{UpdateStates}
\usage{
UpdateStatesNewINIT(Y = Y, trans = qok, .mu, m, SumPerK,
  iniDist = c("PropObs", "UniIni"))
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

