% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Zagg_HMM.R
\name{Zagg_HMM}
\alias{Zagg_HMM}
\title{Processing output of label switching function}
\usage{
Zagg_HMM(USout, .Y = Y)
}
\arguments{
\item{output}{of HMM label switching function, computes parameter estimates and some error values.}
}
\description{
explain here
}
\examples{
#nope
}
\keyword{postprocessing}

