% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Zhmm_PP.R
\name{Zhmm_PP}
\alias{Zhmm_PP}
\title{PostProcessing function univ}
\usage{
Zhmm_PP(run, burn = 1000, prep = 1000, isSim = TRUE, minq4PLOT = 0.05,
  simlabel = "sim")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

