% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Zswitch_hmm}
\alias{Zswitch_hmm}
\title{PostProcessing function univ}
\usage{
Zswitch_hmm(GrunK0, PropMin = 0.1)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

