% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Funkme_MetHast_Q}
\alias{Funkme_MetHast_Q}
\title{Funkme_MetHast_Q}
\usage{
Funkme_MetHast_Q(.q0new, .q0old, .z1)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
text...
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

