% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getq0}
\alias{getq0}
\title{Function to compute stationary distribution of HMM}
\usage{
getq0(Q)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

