% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gibbsHMM_PT.R
\name{gibbsHMM_PT}
\alias{gibbsHMM_PT}
\title{gibbsHMM_PT}
\usage{
gibbsHMM_PT(YZ, M = 2000, K = 5, mu0 = 0, var0 = 100, alphaMAX = 1,
  alphaMin = 1e-05, J = 20, lab = "sim")
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

