% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gibbsHMM_PT_ALLCHAINS}
\alias{gibbsHMM_PT_ALLCHAINS}
\title{gibbsHMM_PT}
\usage{
gibbsHMM_PT_ALLCHAINS(YZ, M = 2000, K = 5, mu0 = 0, var0 = 100,
  alphaMin = 0.5, J = 10)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

