% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotK0_HMM}
\alias{plotK0_HMM}
\title{Function to plot the number of non-empty groups}
\usage{
plotK0_HMM(runResult, burn = 1000, Ptitle = " ",
  PlotName = "nonemptyK0.pdf")
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
...
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

