% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReplicateSimer3}
\alias{ReplicateSimer3}
\title{ReplicateSimer2}
\usage{
ReplicateSimer3(N, n, Kfit = 10, SimID, ITERATIONS, BURN, AMAX, AMIN,
  PRIOR_TYPE, GridSize = 50, PTchain = 30)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
parallel tempering with a column prior - option to mix over column or stick to j=1
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

