% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimHMM2K.R
\name{SimHMM2K}
\alias{SimHMM2K}
\title{Function to compute stationary distribution of HMM}
\usage{
SimHMM2K(q1, q2, mu1, mu2, n = 100)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

