% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unswStates}
\alias{unswStates}
\title{unswStates}
\usage{
unswStates(Out, K = K)
}
\arguments{
\item{mydata,}{alpha, Krange}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#
 #
}
\keyword{bayes}
\keyword{bayesm}
\keyword{factor,}

