% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Zhmm_PostProc}
\alias{Zhmm_PostProc}
\title{PostProcessing function univ}
\usage{
Zhmm_PostProc(Grun, mydata, burn = 1000, Thin = 1, prep = 1000,
  isSim = TRUE, simlabel = "sim")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

