% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ZswitchHMM.R
\name{Zswitch_hmm}
\alias{Zswitch_hmm}
\title{PostProcessing function univ}
\usage{
Zswitch_hmm(GrunK0, PropMin = 0.01)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart distu
}
\examples{
#nope
}
\keyword{Wishart}

